# docs/swagger.yaml
basePath: /
definitions:
...
  models.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      details:
        example: provider字段不能为空
        type: string
      error:
        example: true
        type: boolean
      message:
        example: 参数验证失败
        type: string
    type: object
...
  models.JobListResponse:
    properties:
      count:
        example: 3
        type: integer
      jobs:
        items:
          $ref: '#/definitions/models.JobResponse'
        type: array
    type: object
  models.JobRequest:
    properties:
      cron_expr:
        example: 0 2 * * *
        type: string
      description:
        example: 每日凌晨2点同步火山云账单
        type: string
      enabled:
        example: true
        type: boolean
      name:
        example: daily_volcengine_sync
        type: string
      parameters:
        additionalProperties: true
        type: object
      provider:
        example: volcengine
        type: string
      task_type:
        example: sync
        type: string
    required:
    - cron_expr
    - name
    - provider
    - task_type
    type: object
  models.JobResponse:
    properties:
      created_at:
        example: "2025-09-11T08:13:24Z"
        type: string
      cron_expr:
        example: 0 2 * * *
        type: string
      description:
        example: 每日凌晨2点同步火山云账单
        type: string
      enabled:
        example: true
        type: boolean
      id:
        example: job_123456
        type: string
      last_run:
        example: "2025-09-11T02:00:00Z"
        type: string
      name:
        example: daily_volcengine_sync
        type: string
      next_run:
        example: "2025-09-12T02:00:00Z"
        type: string
      parameters:
        additionalProperties: true
        type: object
      provider:
        example: volcengine
        type: string
      task_type:
        example: sync
        type: string
      updated_at:
        example: "2025-09-11T08:13:24Z"
        type: string
    type: object
  models.MessageResponse:
    properties:
      message:
        example: 操作成功
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  models.SchedulerStatus:
    properties:
      entries:
        example: 3
        type: integer
      job_count:
        example: 3
        type: integer
      running:
        example: true
        type: boolean
      timestamp:
        example: "2025-09-11T08:13:24Z"
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@goscan.com
    name: API Support
  description: |-
    云账单数据同步服务API文档
    支持火山云、阿里云等多云平台的账单数据自动同步到ClickHouse数据库
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Goscan API
  version: "1.0"
paths:
...
  /scheduler/jobs:
    get:
      consumes:
      - application/json
      description: Returns information of all configured scheduled jobs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JobListResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get scheduled job list
      tags:
      - Schedule Management
    post:
      consumes:
      - application/json
      description: Create a new scheduled sync job
      parameters:
      - description: Scheduled job request parameters
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/models.JobRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create scheduled job
      tags:
      - Schedule Management
  /scheduler/jobs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete specified scheduled job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete scheduled job
      tags:
      - Schedule Management
    get:
      consumes:
      - application/json
      description: Returns detailed information of scheduled job by job ID
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.JobResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get specific scheduled job
      tags:
      - Schedule Management
    put:
      consumes:
      - application/json
      description: Update scheduled job configuration for specified ID
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Scheduled job update parameters
        in: body
        name: job
        required: true
        schema:
          $ref: '#/definitions/models.JobRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update scheduled job
      tags:
      - Schedule Management
  /scheduler/jobs/{id}/trigger:
    post:
      consumes:
      - application/json
      description: Immediately execute specified scheduled job without waiting for
        next scheduled time
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Manually trigger scheduled job
      tags:
      - Schedule Management
...
  /scheduler/status:
    get:
      consumes:
      - application/json
      description: Returns running status information of the task scheduler
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SchedulerStatus'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get scheduler status
      tags:
      - Schedule Management
...
schemes:
- http
- https
swagger: "2.0"